+++ zigzag.legendweathers/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 1.1.6
-- Improved the Majora Moon position in the sky for the following moons : EGypt, Noctis and Luigis Mansion
+- Improved the Majora Moon position in the sky for the following moons : EGypt, Noctis, Terra and Luigis Mansion
 
 ## 1.1.5
 - Recompiled for v70